changing float to double

+++ include/util/common.h
@@ -6,7 +6,6 @@
 #include <math.h>
 #include <string.h>
 #include <limits.h>
-#include <float.h>
 
 /* Error messages */
 
@@ -42,7 +41,7 @@ typedef struct timeval timer;
 #endif
 
 int    *AllocIntArray(int n);   /* It allocates 1D array of n integers */
-float  *AllocFloatArray(int n); /* It allocates 1D array of n floats */
+double  *AllocFloatArray(int n); /* It allocates 1D array of n doubles */
 
 void Error(char *msg,char *func); /* It prints error message and exits
                                      the program. */
